# web/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_before_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before Id"
          }
        },
        "required": [
          "items",
          "next_before_id"
        ],
        "title": "JobListOut",
        "type": "object"
      },
      "JobOut": {
        "description": "One ``jobs`` row for the Ops view (#44). ``payload`` is never exposed.",
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dedupe_key": {
            "title": "Dedupe Key",
            "type": "string"
          },
          "error_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Class"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "locked_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locked By"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "run_after": {
            "format": "date-time",
            "title": "Run After",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "video_id": {
            "title": "Video Id",
            "type": "string"
          },
          "video_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Title"
          }
        },
        "required": [
          "id",
          "video_id",
          "video_title",
          "kind",
          "dedupe_key",
          "state",
          "priority",
          "attempts",
          "error_class",
          "last_error",
          "run_after",
          "locked_by",
          "heartbeat_at",
          "finished_at",
          "created_at"
        ],
        "title": "JobOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ytdigest",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ops/jobs": {
      "get": {
        "description": "A page of ``jobs``, newest first (#44). ``payload`` is never exposed.",
        "operationId": "get_ops_jobs_ops_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "running",
                    "done",
                    "dead"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "analyze",
                    "ingest",
                    "transcribe"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "error_class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "PERMANENT_SOURCE",
                    "TRANSIENT_NETWORK",
                    "RATE_LIMITED",
                    "TOOL_FAILURE",
                    "LLM_INVALID_OUTPUT",
                    "LLM_UNAVAILABLE",
                    "RESOURCE",
                    "BUG",
                    "none"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Class"
            }
          },
          {
            "in": "query",
            "name": "before_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9223372036854775807,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ops Jobs"
      }
    },
    "/ops/jobs/{id}/retry": {
      "post": {
        "description": "Revive one ``dead`` job in place: same id, ``attempts`` reset to 0 (#44).\n\nOne conditional ``UPDATE`` (``PostgresQueue.retry_dead``) decides and\nwrites the new state together, so a concurrent change can only ever be\ncaught, never missed. The one log line, ``ops.job_retried``, is emitted\nonly on success; a refusal writes nothing and leaves no trace beyond the\nresponse.",
        "operationId": "post_ops_job_retry_ops_jobs__id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Ops Job Retry"
      }
    },
...
  }
}